int32/int64 dtypes, slice, public broadcastTo (PR15)
- DType gains int32/int64 as runtime tags (Int32Array/BigInt64Array);
  Tensor stays Tensor<S>, gather/scatter stay Tensor<[E]>
- slice folds narrow over a per-axis spec (number | [start,end] | null)
- broadcastTo is public with BroadcastToCheck (expand-only)
- native leaf buffer is now typed bytes; int leaves feed gather/scatter
  indices directly (no f32 16.7M-row limit)
- gnca graphs store src/dst as int32

+++ examples/gnca/checkpoint.ts
@@ -53,8 +53,8 @@ export function checkpointGraph(checkpoint: Checkpoint): {
       dim: checkpoint.dim,
     },
     edges: {
-      src: Float32Array.from(checkpoint.edges.src),
-      dst: Float32Array.from(checkpoint.edges.dst),
+      src: Int32Array.from(checkpoint.edges.src),
+      dst: Int32Array.from(checkpoint.edges.dst),
       count: checkpoint.edges.src.length,
     },
   }
